+++ simple_file_saver/lib/simple_file_saver.dart
@@ -1,91 +1,39 @@
-import 'dart:typed_data';
-
 import 'package:flutter/widgets.dart';
 import 'package:simple_file_saver_platform_interface/simple_file_saver_platform_interface.dart';
 
-SimpleFileSaverPlatform get _platform => SimpleFileSaverPlatform.instance;
+export 'package:simple_file_saver_platform_interface/simple_file_saver_platform_interface.dart';
 
 class SimpleFileSaver {
   SimpleFileSaver._();
 
-  /// Save the file to the default location.
+  /// Save to the default location silently or prompt the user to save as.
   ///
   /// Android defaults to the downloads directory.
   ///
   /// iOS defaults to the documents directory, requires setting the 'Supports Document Browser' key
   /// to YES in the info.plist to reveal the directory to the other apps.
-  static Future<bool> saveFile({
-    required Uint8List dataBytes,
-    required String fileName,
-    String? mimeType,
+  ///
+  /// saveAs has no effect on the web, since it depends on the browser's setting.
+  static Future<String?> saveFile({
+    required FileSaveInfo fileInfo,
+    bool saveAs = false,
   }) {
-    return _platform.saveFile(
-      dataBytes: dataBytes,
-      fileName: fileName,
-      mimeType: mimeType,
+    return SimpleFileSaverPlatform.instance.saveFile(
+      fileInfo: fileInfo,
+      saveAs: saveAs,
     );
   }
 
-  /// Prompt to the users and save the file to the specified location.
-  static Future<bool> saveFileAs({
-    required Uint8List dataBytes,
-    required String fileName,
-    String? mimeType,
-  }) {
-    return _platform.saveFileAs(
-      dataBytes: dataBytes,
-      fileName: fileName,
-      mimeType: mimeType,
-    );
-  }
-
-  /// Download the file by bytes.
-  static Future<void> downloadFileByBytes({
-    required Uint8List dataBytes,
-    String? fileName,
-  }) {
-    return _platform.downloadFileByBytes(
-      dataBytes: dataBytes,
-      fileName: fileName,
-    );
-  }
-
-  /// Download the file by uri.
-  static Future<void> downloadFileByUri({
-    required Uri uri,
-    String? fileName,
-  }) {
-    return _platform.downloadFileByUri(
-      uri: uri,
-      fileName: fileName,
-    );
-  }
-
-  /// Build a Link widget for the web only, and download the file by bytes
-  static Widget webOnlyBytesDownloadLinkBuilder({
-    required Uint8List dataBytes,
-    String? fileName,
+  /// Build a Link widget to enable the right click option 'save link as' on the web platform
+  ///
+  /// This method only implemented on the web platform
+  static Widget downloadLinkBuilder({
+    required FileSaveInfo fileInfo,
     LinkTarget target = LinkTarget.blank,
     required DownloadLinkBuilder builder,
   }) {
-    return _platform.webOnlyBytesDownloadLinkBuilder(
-      dataBytes: dataBytes,
-      fileName: fileName,
-      target: target,
-      builder: builder,
-    );
-  }
-
-  /// Build a Link widget for the web only, and download the file by Uri
-  static Widget webOnlyUriDownloadLinkBuilder({
-    required Uri uri,
-    String? fileName,
-    LinkTarget target = LinkTarget.blank,
-    required DownloadLinkBuilder builder,
-  }) {
-    return _platform.webOnlyUriDownloadLinkBuilder(
-      uri: uri,
-      fileName: fileName,
+    return SimpleFileSaverPlatform.instance.downloadLinkBuilder(
+      fileInfo: fileInfo,
       target: target,
       builder: builder,
     );