+++ simple_file_saver/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+
+* BREAKING CHANGE: The API in now simplified into one single method `saveFile()` and an additional method `downloadLinkBuilder()` only for the web platform.
+* The new API takes a `FileSaveInfo` object which you may create it from either `FileSaveInfo.fromBytes()` or `FileSaveInfo.fromUrl()`.
+
 ## 1.1.0
 
 * Add web implementation to the plugin.